thread_local

Module thread_id

Structs

Thread 🔒

Data which is unique to the current thread while it is running. A thread ID may be reused after a thread exits.

ThreadGuard 🔒

ThreadIdManager 🔒

Thread ID manager which allocates thread IDs. It attempts to aggressively reuse thread IDs where possible to avoid cases where a ThreadLocal grows indefinitely when it is used by many short-lived threads.

Constants

THREAD 🔒

THREAD_GUARD 🔒

Statics

THREAD_ID_MANAGER 🔒

Functions

get 🔒

Returns a thread ID for the current thread, allocating one if needed.

get_slow 🔒

Out-of-line slow path for allocating a thread ID.

try_get 🔒

Returns a thread ID for the current thread, not allocating one if needed. This avoids registering a thread-local destructor.