Struct thread_local::ThreadLocal

pub struct ThreadLocal<T: Send> { /* fields omitted */ }

Thread-local variable wrapper

See the module-level documentation for more.

Methods

impl<T: Send> ThreadLocal<T>

pub fn new() -> ThreadLocal<T>

Creates a new empty ThreadLocal.

pub fn get(&self) -> Option<&T>

Returns the element for the current thread, if it exists.

pub fn get_or<F>(&self, create: F) -> &T where
F: FnOnce() -> T,

Returns the element for the current thread, or creates it if it doesn't exist.

pub fn get_or_try<F, E>(&self, create: F) -> Result<&T, E> where
F: FnOnce() -> Result<T, E>,

Returns the element for the current thread, or creates it if it doesn't exist. If create fails, that error is returned and no element is added.

Important traits for IterMut<'a, T>
impl<'a, T: Send + 'a> Iterator for IterMut<'a, T> type Item = &'a mut T;

pub fn iter_mut(&mut self) -> IterMut<T>

Returns a mutable iterator over the local values of all threads.

Since this call borrows the ThreadLocal mutably, this operation can be done safely---the mutable borrow statically guarantees no other threads are currently accessing their associated values.

pub fn clear(&mut self)

Removes all thread-specific values from the ThreadLocal, effectively reseting it to its original state.

Since this call borrows the ThreadLocal mutably, this operation can be done safely---the mutable borrow statically guarantees no other threads are currently accessing their associated values.

impl<T: Send + Default> ThreadLocal<T>

pub fn get_or_default(&self) -> &T

Returns the element for the current thread, or creates a default one if it doesn't exist.