Struct thread_local::ThreadLocal

pub struct ThreadLocal<T: Send> { /* private fields */ }

Thread-local variable wrapper

See the module-level documentation for more.

Implementations

impl<T: Send> ThreadLocal<T>

pub fn new() -> ThreadLocal<T>

Creates a new empty ThreadLocal.

pub fn with_capacity(capacity: usize) -> ThreadLocal<T>

Creates a new ThreadLocal with an initial capacity. If less than the capacity threads access the thread local it will never reallocate. The capacity may be rounded up to the nearest power of two.

pub fn get(&self) -> Option<&T>

Returns the element for the current thread, if it exists.

pub fn get_or<F>(&self, create: F) -> &Twhere
F: FnOnce() -> T,

Returns the element for the current thread, or creates it if it doesn’t exist.

pub fn get_or_try<F, E>(&self, create: F) -> Result<&T, E>where
F: FnOnce() -> Result<T, E>,

Returns the element for the current thread, or creates it if it doesn’t exist. If create fails, that error is returned and no element is added.

pub fn iter(&self) -> Iter<'_, T>where
T: Sync,

Returns an iterator over the local values of all threads in unspecified order.

This call can be done safely, as T is required to implement Sync.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Returns a mutable iterator over the local values of all threads in unspecified order.

Since this call borrows the ThreadLocal mutably, this operation can be done safely—the mutable borrow statically guarantees no other threads are currently accessing their associated values.

pub fn clear(&mut self)

Removes all thread-specific values from the ThreadLocal, effectively reseting it to its original state.

Since this call borrows the ThreadLocal mutably, this operation can be done safely—the mutable borrow statically guarantees no other threads are currently accessing their associated values.

impl<T: Send + Default> ThreadLocal<T>

pub fn get_or_default(&self) -> &T

Returns the element for the current thread, or creates a default one if it doesn’t exist.